perf: optimize server performance and clean up code
- Replace body-parser with express.json()
- Add static file caching (1 day)
- Disable x-powered-by header
- Fix async Promise anti-pattern
- Use ES6 shorthand syntax
- Extract HTML content to public/index.html
- Move styles to public/css/style.css
- Move client JS to public/js/script.js
- Configure Express static file serving
