+++ package.json
@@ -11,10 +11,10 @@
   "license": "ISC",
   "type": "module",
   "dependencies": {
-    "body-parser": "^2.2.1",
+    "body-parser": "^2.2.2",
     "express": "^5.2.1",
     "input": "^1.0.1",
-    "open": "8.4.2",
+    "open": "^11.0.0",
     "telegram": "^2.26.22"
   }
 }
