+++ public/css/style.css
@@ -1,0 +1,131 @@
+:root {
+    --bg: #0f172a;
+    --card: #1e293b;
+    --input: #334155;
+    --text: #f8fafc;
+    --muted: #94a3b8;
+    --accent: #3b82f6;
+    --accent-hover: #2563eb;
+    --success: #10b981;
+    --error: #ef4444;
+    --font-fa: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+    --font-en: 'Inter', system-ui, -apple-system, sans-serif;
+}
+
+body {
+    font-family: var(--font-fa);
+    background-color: var(--bg);
+    color: var(--text);
+    margin: 0;
+    padding: 20px;
+    display: flex;
+    justify-content: center;
+    min-height: 100vh;
+    transition: all 0.3s ease;
+}
+
+html[dir="ltr"] body { font-family: var(--font-en); }
+html[dir="ltr"] .debug-console { text-align: left; }
+
+.container {
+    width: 100%;
+    max-width: 1100px;
+    background-color: var(--card);
+    padding: 30px;
+    border-radius: 16px;
+    box-shadow: 0 20px 50px rgba(0,0,0,0.5);
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    position: relative;
+}
+
+.header-controls {
+    position: absolute;
+    top: 25px;
+    left: 25px;
+    z-index: 10;
+}
+html[dir="ltr"] .header-controls {
+    left: auto;
+    right: 25px;
+}
+
+select.lang-select {
+    background-color: var(--input);
+    color: var(--text);
+    border: 1px solid #475569;
+    padding: 8px 12px;
+    border-radius: 8px;
+    font-size: 13px;
+    cursor: pointer;
+    outline: none;
+    font-family: inherit;
+    transition: 0.2s;
+    appearance: none;
+    -webkit-appearance: none;
+    background-image: url("data:image/svg+xml;charset=US-ASCII,%3Csvg%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20width%3D%22292.4%22%20height%3D%22292.4%22%3E%3Cpath%20fill%3D%22%23cbd5e1%22%20d%3D%22M287%2069.4a17.6%2017.6%200%200%200-13-5.4H18.4c-5%200-9.3%201.8-12.9%205.4A17.6%2017.6%200%200%200%200%2082.2c0%205%201.8%209.3%205.4%2012.9l128%20127.9c3.6%203.6%207.8%205.4%2012.8%205.4s9.2-1.8%2012.8-5.4L287%2095c3.5-3.5%205.4-7.8%205.4-12.8%200-5-1.9-9.2-5.5-12.8z%22%2F%3E%3C%2Fsvg%3E");
+    background-repeat: no-repeat;
+    background-size: 10px;
+    background-position: left 10px center;
+    padding-left: 30px;
+    padding-right: 12px;
+}
+
+html[dir="ltr"] select.lang-select {
+    background-position: right 10px center;
+    padding-right: 30px;
+    padding-left: 12px;
+}
+
+select.lang-select:hover {
+    border-color: var(--accent);
+    background-color: #3f4a5e;
+}
+
+h1 { text-align: center; color: var(--accent); margin: 0; margin-top: 5px; }
+p.sub { text-align: center; color: var(--muted); margin: 5px 0 20px 0; font-size: 14px; }
+
+.progress-box { background: var(--input); height: 8px; border-radius: 4px; overflow: hidden; }
+.progress-bar { height: 100%; width: 0%; background: linear-gradient(90deg, var(--accent), #60a5fa); transition: width 0.3s ease; }
+.status-text { text-align: center; font-size: 13px; color: var(--muted); margin-top: 5px; direction: ltr; }
+
+.grid { display: grid; grid-template-columns: 1fr 1fr; gap: 20px; }
+@media(max-width: 768px) { .grid { grid-template-columns: 1fr; } }
+
+label { display: block; margin-bottom: 8px; font-weight: bold; font-size: 14px; }
+
+textarea {
+    width: 100%; height: 350px; background-color: var(--input); color: #e2e8f0;
+    border: 1px solid #475569; border-radius: 10px; padding: 15px;
+    font-family: monospace; font-size: 11px; resize: vertical; outline: none;
+    line-height: 1.5; box-sizing: border-box;
+}
+textarea:focus { border-color: var(--accent); }
+
+button {
+    width: 100%; padding: 14px; margin-top: 10px; border: none; border-radius: 10px;
+    font-weight: bold; font-size: 15px; cursor: pointer; transition: 0.2s; color: white; font-family: inherit;
+}
+
+.btn-start { background-color: var(--accent); }
+.btn-start:hover { background-color: var(--accent-hover); }
+.btn-start:disabled { background-color: #475569; cursor: not-allowed; opacity: 0.7; }
+
+.btn-copy { background-color: var(--success); }
+.btn-copy:hover { background-color: #059669; }
+
+.debug-console {
+    background: black; color: #00ff00; font-family: monospace; font-size: 11px;
+    padding: 10px; height: 80px; overflow-y: auto; border-radius: 8px;
+    border: 1px solid #333; direction: ltr; text-align: left; margin-top: 10px;
+}
+.error-log { color: #ff5555; }
+
+.toast {
+    visibility: hidden; min-width: 250px; background-color: #333; color: #fff;
+    text-align: center; border-radius: 8px; padding: 12px; position: fixed;
+    z-index: 99; left: 50%; bottom: 30px; transform: translateX(-50%);
+    box-shadow: 0 5px 15px rgba(0,0,0,0.3); opacity: 0; transition: opacity 0.3s, bottom 0.3s;
+}
+.toast.show { visibility: visible; opacity: 1; bottom: 50px; }
